quantity change type input to int

+++ src/components/inventory/RecordSaleModal.tsx
@@ -221,7 +221,7 @@ const RecordSaleModal: React.FC<RecordSaleModalProps> = ({
                 type="number"
                 required
                 value={formData.quantity}
-                onChange={(e) => handleInputChange('quantity', parseFloat(e.target.value))}
+                onChange={(e) => handleInputChange('quantity', parseInt(e.target.value))}
                 className="block w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-green-500 focus:border-transparent"
                 placeholder="Quantity sold"
               />
